requirements.txt
# RSS 解析
feedparser==6.0.11

# HTTP 请求
requests==2.31.0

# 命令行工具
click==8.1.7
rich==13.7.1

# 环境变量
python-dotenv==1.0.1

# AI SDK
openai==1.30.1

# Token 计数
tiktoken==0.7.0

# 文本处理
beautifulsoup4==4.12.3
lxml==5.2.1

# 配置文件解析
pyyaml==6.0.1
